fix(package): try better package name replace
From my local test it should have worked as it was but in fails on GitHub actions: https://github.com/pollination/honeybee-radiance/pull/22/checks?check_run_id=1832715199#step:5:28

It might also be because of the fact that it tries to load the module before it being installed. We will know when the tests run next.